drop all PixelExtensions usages in product code,
keep in test code for now.
